Updated Qt5 install scripts. Updated the FFTW to look in lib64
Installing Qt5 on Linux is still problematic. You need a valid XServer connection
to allow the GUI installer to run. There is a message that prints out that states
the executable to run in case the Qt5 installer does not execute.

FFTW is now built using CMake instead of the auto-conf based system, at least on linux
For Windows, one still needs to download the prebuilt binaries for FFTW. Should
investigate using Intel Math Kernel Library install as it has an FFTW compatible
interface.

Signed-off-by: Michael Jackson <mjackson@keightie.materials.cmu.edu>